precommit fixed meterology.py
