+++ docs/internal_api/index.rst
@@ -11,33 +11,39 @@ GeoCAT-comp routines
    :nosignatures:
    :toctree: ./generated/
 
-   geocat.comp.climatology._contains_datetime_like_objects
+   geocat.comp.climatologies._contains_datetime_like_objects
 
-   geocat.comp.climatology._calculate_center_of_time_bounds
+   geocat.comp.climatologies._calculate_center_of_time_bounds
 
-   geocat.comp.climatology._find_time_invariant_vars
+   geocat.comp.climatologies._find_time_invariant_vars
 
-   geocat.comp.climatology._get_time_coordinate_info
+   geocat.comp.climatologies._get_time_coordinate_info
 
-   geocat.comp.climatology._setup_clim_anom_input
+   geocat.comp.climatologies._setup_clim_anom_input
 
-   geocat.comp.climatology._validate_freq
+   geocat.comp.climatologies._validate_freq
 
    geocat.comp.comp_util._is_duck_array
 
-   geocat.comp.dewtemp._dewtemp
-
    geocat.comp.eofunc._generate_eofs_solver
 
-   geocat.comp.heat_index._nws_eqn
+   geocat.comp.interpolation._pressure_from_hybrid
 
-   geocat.comp.interp_hybrid_to_pressure._pressure_from_hybrid
+   geocat.comp.interpolation._vertical_remap
 
-   geocat.comp.interp_hybrid_to_pressure._vertical_remap
+   geocat.comp.meteorology._dewtemp
+
+   geocat.comp.meteorology._heat_index
+
+   geocat.comp.meteorology._nws_eqn
+
+   geocat.comp.meteorology._relhum
+
+   geocat.comp.meteorology._xheat_index
+
+   geocat.comp.meteorology._xrelhum
 
    geocat.comp.polynomial._get_missing_value
-
-   geocat.comp.heat_index._heat_index
 
    geocat.comp.polynomial._unchunk_ifneeded
 
@@ -54,9 +60,3 @@ GeoCAT-comp routines
    geocat.comp.polynomial._to_numpy_ndarray
 
    geocat.comp.polynomial._ndpolyval
-
-   geocat.comp.relhum._relhum
-
-   geocat.comp.heat_index._xheat_index
-
-   geocat.comp.relhum._xrelhum
