+++ src/geocat/comp/climatologies.py
@@ -106,7 +106,7 @@ def _infer_calendar_name(dates):
     This code was taken from `xarray/coding/times.py <https://github.com/pydata/xarray/blob/75eefb8e95a70f1667623a8418d81da3f3148a40/xarray/coding/times.py>_`
     as the function is considered internal by xarray and could change at
     anytime. It was copied to preserve the version that is compatible with
-    functions in climatology.py
+    functions in climatologies.py
     """
     if np.asarray(dates).dtype == "datetime64[ns]":
         return "proleptic_gregorian"
